requirements.txt
selenium
pytest
webdriver-manager
pytest-xdist
allure-pytest